Extracted input reading into another class
